+++ pioemu/conditions.py
@@ -11,47 +11,51 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-def clock_cycles_reached(target_value):
+from typing import Callable, Literal
+from .state import State
+
+
+def clock_cycles_reached(target_value: int) -> Callable[[int, State], bool]:
     return lambda _, state: state.clock >= target_value
 
 
-def always(_):
+def always(_: State) -> Literal[True]:
     return True
 
 
-def gpio_low(pin_number, state):
+def gpio_low(pin_number: int, state: State) -> bool:
     return not gpio_high(pin_number, state)
 
 
-def gpio_high(pin_number, state):
-    return state.pin_values & (1 << pin_number)
+def gpio_high(pin_number: int, state: State) -> bool:
+    return state.pin_values & (1 << pin_number) != 0
 
 
-def transmit_fifo_not_empty(state):
+def transmit_fifo_not_empty(state: State) -> bool:
     return len(state.transmit_fifo) > 0
 
 def receive_fifo_not_full(state):
     return len(state.receive_fifo) < 32
 
-def x_register_equals_zero(state):
+def x_register_equals_zero(state: State) -> bool:
     return state.x_register == 0
 
 
-def x_register_not_equal_to_zero(state):
+def x_register_not_equal_to_zero(state: State) -> bool:
     return state.x_register != 0
 
 
-def y_register_equals_zero(state):
+def y_register_equals_zero(state: State) -> bool:
     return state.y_register == 0
 
 
-def y_register_not_equal_to_zero(state):
+def y_register_not_equal_to_zero(state: State) -> bool:
     return state.y_register != 0
 
 
-def x_register_not_equal_to_y_register(state):
+def x_register_not_equal_to_y_register(state: State) -> bool:
     return state.x_register != state.y_register
 
 
-def output_shift_register_not_empty(state):
+def output_shift_register_not_empty(state: State) -> bool:
     return state.output_shift_register.counter != 32