+++ CHANGELOG.md
@@ -1,16 +1,19 @@
 # Changelog
 
+## Unreleased
+
+### Added
+- Support for the OUT PC instruction. Thanks @[aaronjamt](https://github.com/aaronjamt).
+- Type-hints for documentation and to improve maintainability
+
 ## v0.78.0 (2023-03-25)
 
 ### Added
-
 - Support for the IN instruction.
 - Support for the MOV PC instruction. Thanks @[aaronjamt](https://github.com/aaronjamt).
 
 ### Changed
-
 - Improved support for advancing the program counter.
 
 ### Removed
-
 - Dropped support for older versions of Python (< 3.10.10)